app/server/pyproject.toml
[project]
name = "valargen-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.45",
    "alembic>=1.17.2",
    "passlib[argon2]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "authlib>=1.6.6",
    "redis>=7.1.0",
    "pydantic-settings>=2.12.0",
    "email-validator>=2.0.0",
    "httpx[socks]>=0.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
